Remove primitives
