+++ meson.build
@@ -55,7 +55,6 @@ libsearch = static_library(
         SRC_DIR / 'search' / 'LocalSearch.cpp',
         SRC_DIR / 'search' / 'PerturbationManager.cpp',
         SRC_DIR / 'search' / 'Route.cpp',
-        SRC_DIR / 'search' / 'primitives.cpp',
         SRC_DIR / 'search' / 'RelocateWithDepot.cpp',
         SRC_DIR / 'search' / 'RemoveAdjacentDepot.cpp',
         SRC_DIR / 'search' / 'RemoveOptional.cpp',
