+++ pyvrp/cpp/search/Exchange.h
@@ -3,7 +3,6 @@
 
 #include "LocalSearchOperator.h"
 #include "Route.h"
-#include "primitives.h"
 
 #include <cassert>
