+++ pyvrp/cpp/search/PerturbationManager.cpp
@@ -1,5 +1,4 @@
 #include "PerturbationManager.h"
-#include "primitives.h"
 
 #include <cassert>
 #include <stdexcept>
