+++ pyvrp/cpp/search/RemoveAdjacentDepot.cpp
@@ -1,7 +1,5 @@
 #include "RemoveAdjacentDepot.h"
 
-#include "primitives.h"
-
 #include <cassert>
 #include <limits>
 
@@ -17,11 +15,16 @@ RemoveAdjacentDepot::evaluate(Route::Node *U,
     if (!U->route())
         return std::make_pair(0, false);
 
+    auto *route = U->route();
     Cost bestCost = std::numeric_limits<Cost>::max();
 
     if (p(U)->isReloadDepot())
     {
-        Cost deltaCost = removeCost(p(U), data, costEvaluator);
+        Cost deltaCost = 0;
+        costEvaluator.deltaCost(deltaCost,
+                                Route::Proposal(route->before(U->idx() - 2),
+                                                route->after(U->idx())));
+
         if (deltaCost < bestCost)
         {
             bestCost = deltaCost;
@@ -31,7 +34,11 @@ RemoveAdjacentDepot::evaluate(Route::Node *U,
 
     if (n(U)->isReloadDepot())
     {
-        Cost deltaCost = removeCost(n(U), data, costEvaluator);
+        Cost deltaCost = 0;
+        costEvaluator.deltaCost(deltaCost,
+                                Route::Proposal(route->before(U->idx()),
+                                                route->after(U->idx() + 2)));
+
         if (deltaCost < bestCost)
         {
             bestCost = deltaCost;
