+++ pyvrp/cpp/search/RemoveOptional.cpp
@@ -1,12 +1,13 @@
 #include "RemoveOptional.h"
 
-#include "primitives.h"
+#include <cassert>
 
 using pyvrp::search::RemoveOptional;
 
 std::pair<pyvrp::Cost, bool>
 RemoveOptional::evaluate(Route::Node *U, CostEvaluator const &costEvaluator)
 {
+    assert(U->client());
     stats_.numEvaluations++;
 
     ProblemData::Client const &uData = data.location(U->client());
@@ -26,7 +27,15 @@ RemoveOptional::evaluate(Route::Node *U, CostEvaluator const &costEvaluator)
             return std::make_pair(0, false);  // group member
     }
 
-    auto const deltaCost = removeCost(U, data, costEvaluator);
+    auto *route = U->route();
+    Cost deltaCost
+        = uData.prize
+          - Cost(route->numClients() == 1) * route->fixedVehicleCost();
+
+    costEvaluator.deltaCost(deltaCost,
+                            Route::Proposal(route->before(U->idx() - 1),
+                                            route->after(U->idx() + 1)));
+
     return std::make_pair(deltaCost, deltaCost < 0);
 }
