+++ pyvrp/search/_search.pyi
@@ -225,13 +225,3 @@ class Node:
     def is_start_depot(self) -> bool: ...
     def is_end_depot(self) -> bool: ...
     def is_reload_depot(self) -> bool: ...
-
-def insert_cost(
-    U: Node, V: Node, data: ProblemData, cost_evaluator: CostEvaluator
-) -> int: ...
-def inplace_cost(
-    U: Node, V: Node, data: ProblemData, cost_evaluator: CostEvaluator
-) -> int: ...
-def remove_cost(
-    U: Node, data: ProblemData, cost_evaluator: CostEvaluator
-) -> int: ...
